# Optimize ASCII character width calculation